added almost complete project
removed older version of project and added the version that is essentially complete, aside from implementing Junit
