+++ sample/Main.java
@@ -14,7 +14,7 @@ public class Main extends Application {
     public void start(Stage primaryStage) throws Exception{
         Parent root = FXMLLoader.load(getClass().getResource("sample.fxml"));
         primaryStage.setTitle("Algorithm Tester");
-        primaryStage.setScene(new Scene(root, 700, 375));
+        primaryStage.setScene(new Scene(root, 620, 375));
         primaryStage.show();
     }
